requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0

# Data sources and APIs
yfinance>=0.1.70
yahooquery>=2.3.0
fredapi>=0.5.0
requests>=2.25.0
certifi>=2023.7.22
python-dotenv>=0.19.0
pandas-datareader==0.10.0

# Data analysis
statsmodels==0.13.5
scikit-learn==1.2.2
pyportfolioopt==1.5.5

# Visualization
plotly==5.18.0
ipywidgets>=7.6.0

# Financial calculation
empyrical==0.5.5
arch==5.6.0

# Dashboard and reporting
tabpy==2.13.0
tableauserverclient==0.19.0
streamlit==1.27.0

# Utilities
tqdm==4.66.1
wheel==0.40.0
jupyter>=1.0.0
